Renamed and moved classes
- Renamed inputcontainer package to playback
- Renamed InputContainer to PlaybackController
- Moved ContainerSerialiser from util to package playback
- Renamed ContainerSerialiser to PlaybackSerialiser
- Renamed ContainerStateClient to TASstateClient
- Renamed ContainerStateServer to TASstateServer
- Moved ShieldDownloader from shield to package util and removed shield
- Moved TASstate to inner class of PlaybackController
- Moved TickInputContainer to inner class of PlaybackController
