+++ src/main/java/com/minecrafttas/tasmod/commands/fullrecord/CommandFullRecord.java
@@ -1,7 +1,7 @@
 package com.minecrafttas.tasmod.commands.fullrecord;
 
 import com.minecrafttas.tasmod.TASmod;
-import com.minecrafttas.tasmod.inputcontainer.TASstate;
+import com.minecrafttas.tasmod.playback.PlaybackController.TASstate;
 import com.minecrafttas.tasmod.savestates.server.SavestateState;
 import com.minecrafttas.tasmod.savestates.server.exceptions.SavestateException;
