+++ src/main/java/com/minecrafttas/tasmod/mixin/shields/MixinRenderItem.java
@@ -5,7 +5,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.minecrafttas.tasmod.shield.ShieldDownloader;
+import com.minecrafttas.tasmod.util.ShieldDownloader;
 
 import net.minecraft.client.renderer.RenderItem;
 import net.minecraft.client.renderer.block.model.ItemCameraTransforms;
